\name{hgu95av2frmavecs}
\docType{data}
\alias{hgu95av2frmavecs}
\title{Vectors used by fRMA for hgu95av2.}
\description{This data object was automatically created by the package frmaTools version 1.8.0.}
\usage{data(hgu95av2frmavecs)}
\format{A list with 6 elements.
  \tabular{ll}{
    normVec \tab normalization vector \cr
    probeVec \tab probe effect vector \cr
    probeVarWithin \tab within batch probe variance \cr
    probeVarBetween \tab between batch probe variance \cr
    probesetSD \tab within probeset standard deviation \cr
    medianSE \tab median standard errors \cr
}}

\keyword{datasets}
\examples{
 data(hgu95av2frmavecs) 
 str(hgu95av2frmavecs)
}
